requirements.txt
flask
pymongo
python-dotenv
werkzeug
requests
gunicorn